codec meta single class
