+++ src/main/java/org/jcodec/api/SequenceEncoder.java
@@ -66,7 +66,7 @@ public class SequenceEncoder {
         }
         if (outTrack == null) {
             // Add video track to muxer
-            outTrack = muxer.addVideoTrack(Codec.H264, new VideoCodecMeta(new Size(pic.getWidth(), pic.getHeight()), pic.getColor()));
+            outTrack = muxer.addVideoTrack(Codec.H264, org.jcodec.common.VideoCodecMeta.createSimpleVideoCodecMeta(new Size(pic.getWidth(), pic.getHeight()), pic.getColor()));
         }
 
         // Perform conversion
