+++ src/main/java/org/jcodec/codecs/mpeg12/MPEGDecoder.java
@@ -771,6 +771,6 @@ public class MPEGDecoder extends VideoDecoder {
     public VideoCodecMeta getCodecMeta(ByteBuffer data) {
         ByteBuffer codecPrivate = getSequenceHeader(data.duplicate());
         SequenceHeader sh = SequenceHeader.read(codecPrivate.duplicate());
-        return new VideoCodecMeta(new Size(sh.horizontal_size, sh.vertical_size), ColorSpace.YUV420);
+        return org.jcodec.common.VideoCodecMeta.createSimpleVideoCodecMeta(new Size(sh.horizontal_size, sh.vertical_size), ColorSpace.YUV420);
     }
 }
