+++ src/main/java/org/jcodec/movtool/streaming/tracks/DownmixHelper.java
@@ -21,11 +21,11 @@ import static org.jcodec.common.model.Label.RightSurround;
 import static org.jcodec.common.model.Label.RightTotal;
 import static org.jcodec.common.model.Label.Unused;
 
+import org.jcodec.common.AudioCodecMeta;
 import org.jcodec.common.IntArrayList;
 import org.jcodec.common.io.NIOUtils;
 import org.jcodec.common.logging.Logger;
 import org.jcodec.common.model.Label;
-import org.jcodec.movtool.streaming.AudioCodecMeta;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
