+++ src/main/java/org/jcodec/movtool/streaming/tracks/EditedFramesTrack.java
@@ -3,7 +3,7 @@ import java.lang.IllegalStateException;
 import java.lang.System;
 import java.lang.IllegalArgumentException;
 
-import org.jcodec.movtool.streaming.CodecMeta;
+import org.jcodec.common.CodecMeta;
 import org.jcodec.movtool.streaming.VirtualPacket;
 import org.jcodec.movtool.streaming.VirtualTrack;
