+++ src/main/java/org/jcodec/movtool/streaming/tracks/EditedPCMTrack.java
@@ -2,10 +2,9 @@ package org.jcodec.movtool.streaming.tracks;
 import java.lang.IllegalStateException;
 import java.lang.System;
 
-
+import org.jcodec.common.AudioCodecMeta;
+import org.jcodec.common.CodecMeta;
 import org.jcodec.common.io.NIOUtils;
-import org.jcodec.movtool.streaming.AudioCodecMeta;
-import org.jcodec.movtool.streaming.CodecMeta;
 import org.jcodec.movtool.streaming.VirtualPacket;
 import org.jcodec.movtool.streaming.VirtualTrack;
