+++ src/main/java/org/jcodec/movtool/streaming/tracks/Jpeg2AVCTrack.java
@@ -7,9 +7,9 @@ import java.lang.IllegalArgumentException;
 import org.jcodec.codecs.mjpeg.JpegDecoder;
 import org.jcodec.codecs.mjpeg.JpegToThumb2x2;
 import org.jcodec.codecs.mjpeg.JpegToThumb4x4;
+import org.jcodec.common.VideoCodecMeta;
 import org.jcodec.common.VideoDecoder;
 import org.jcodec.common.model.Size;
-import org.jcodec.movtool.streaming.VideoCodecMeta;
 import org.jcodec.movtool.streaming.VirtualTrack;
 
 /**
