+++ src/main/java/org/jcodec/movtool/streaming/tracks/StereoDownmixTrack.java
@@ -3,10 +3,10 @@ import java.lang.IllegalStateException;
 import java.lang.System;
 import java.lang.IllegalArgumentException;
 
+import org.jcodec.common.AudioCodecMeta;
 import org.jcodec.common.AudioFormat;
+import org.jcodec.common.CodecMeta;
 import org.jcodec.common.model.Label;
-import org.jcodec.movtool.streaming.AudioCodecMeta;
-import org.jcodec.movtool.streaming.CodecMeta;
 import org.jcodec.movtool.streaming.VirtualPacket;
 import org.jcodec.movtool.streaming.VirtualTrack;
